+++ src/components/tables/CategoriesTable.js
@@ -1,5 +1,6 @@
 import moment from 'moment';
 import { useDispatch } from 'react-redux';
+import Swal from 'sweetalert2';
 import TableLoader from '../loaders/TableLoader';
 
 const CategoriesTable = ({ isLoading, categories, deleteHandlers, setEditing }) => {
@@ -7,7 +8,25 @@ const CategoriesTable = ({ isLoading, categories, deleteHandlers, setEditing }) 
 
     // delete category handler
     const deleteCategoryHandler = (id) => {
-        deleteCategory(id);
+        Swal.fire({
+            title: 'Are you sure?',
+            text: "You won't be able to revert this!",
+            icon: 'warning',
+            showCancelButton: true,
+            confirmButtonColor: '#3085d6',
+            cancelButtonColor: '#d33',
+            confirmButtonText: 'Yes, delete it!'
+        }).then((result) => {
+            if (result.isConfirmed) {
+                deleteCategory(id);
+                Swal.fire(
+                    'Deleted!',
+                    'Your file has been deleted.',
+                    'success'
+                )
+            }
+        })
+        
     }
 
     const dispatch = useDispatch();
